backend/python/requirements.txt
pdfplumber
pytesseract
Pillow
pydantic
